Moved legion select box to own component
