+++ src/app/App.css
@@ -36,6 +36,7 @@ footer {
     left: 0;
     width: 100%;
     height: 100%;
+    background-size: cover;
 }
 
 .tooltip-parent:hover .tooltip-data {
